AJ Follow standard JVM package structure conventions
