--- src/main/scala/model/WrappedLocation.scala
@@ -1,30 +1,0 @@
-/**
- *
- * Copyright (c) 2015-present, Total Location Test Paragraph.
- * All rights reserved.
- *
- * This file is part of Where@. Where@ is free software:
- * you can redistribute it and/or modify it under the terms of
- * the GNU General Public License (GPL), either version 3
- * of the License, or (at your option) any later version.
- *
- * Where@ is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. For more details,
- * see the full license at <http://www.gnu.org/licenses/gpl-3.0.en.html>
- *
- */
-
-package model
-
-import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
-import spray.json.DefaultJsonProtocol
-
-
-case class WrappedLocation(lastPing: Long, location: Location)
-
-object WrappedLocation
-
-trait WrappedLocationJsonProtocol extends LocationJsonProtocol with DefaultJsonProtocol with SprayJsonSupport {
-  implicit val wrappedLocationFormat = jsonFormat2(WrappedLocation.apply)
-}
