--- src/test/scala/model/MessageSpec.scala
@@ -1,44 +1,0 @@
-/**
- *
- * Copyright (c) 2015-present, Total Location Test Paragraph.
- * All rights reserved.
- *
- * This file is part of Where@. Where@ is free software:
- * you can redistribute it and/or modify it under the terms of
- * the GNU General Public License (GPL), either version 3
- * of the License, or (at your option) any later version.
- *
- * Where@ is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. For more details,
- * see the full license at <http://www.gnu.org/licenses/gpl-3.0.en.html>
- *
- */
-
-package model
-
-import org.scalatest.{Matchers, WordSpec}
-import spray.json._
-
-/**
- * Author: @aguestuser
- * Date: 9/7/15
- * License: GPLv2 (https://www.gnu.org/licenses/gpl-2.0.html)
- */
-class MessageSpec
-  extends WordSpec
-  with Matchers
-  with JsonProtocols {
-
-  "Message" should {
-
-    "serialize to JSON" in {
-      Message("hi!").toJson.toString() shouldEqual """{"msg":"hi!"}"""
-    }
-
-    "deserialize from JSON" in {
-
-      """{"msg":"hi!"}""".parseJson.convertTo[Message] shouldEqual Message("hi!")
-    }
-  }
-}
